IF_RELENG_7 185198:185562
Former-commit-id: b6412fa95f1e66c64a5d877c43f664cc74534c94

+++ sys/ia64/include/atomic.h
@@ -41,7 +41,7 @@ parameter_list|,
 name|ret
 parameter_list|)
 define|\
-value|__asm __volatile (						\ 		"mov ar.ccv=%2;;\n\t"					\ 		"cmpxchg" #sz "." #sem " %0=%4,%3,ar.ccv\n\t"		\ 		: "=r" (ret), "=m" (*p)					\ 		: "r" (cmpval), "r" (newval), "m" (*p)			\ 		: "memory")
+value|__asm __volatile (						\ 		"mov ar.ccv=%2;;\n\t"					\ 		"cmpxchg" #sz "." #sem " %0=%4,%3,ar.ccv\n\t"		\ 		: "=r" (ret), "=m" (*p)					\ 		: "r" ((uint64_t)cmpval), "r" (newval), "m" (*p)	\ 		: "memory")
 end_define
 
 begin_comment
